Add hostname to a configurable field name
Inspired by fluentd in_forward's source_hostname_key, but done on the endpoint
so that in_forward doesn't have to split apart and re-contruct messagepack
record sets, and so the hostname is used instead of an IP address.

This is useful when using Fluentd to send to a centralized logging system like
Graylog that requires a hostname entry for logs.

Note, format="None" logs use a custom Fluent message structure and do not use a
RecordModifier so hostnames are not added there.

Signed-off-by: Eric Searcy <emsearcy@gmail.com>